TDFold_env_requirements.txt
absl-py==2.1.0
accelerate==0.30.0
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
axial_positional_embedding==0.3.12
biopython==1.79
Brotli==1.0.9
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cramjam==2.8.3
datasets==2.19.0
dgl==0.9.1.post1
diffusers==0.27.2
dill==0.3.8
docker-pycreds==0.4.0
einops==0.8.1
fastparquet==2024.5.0
filelock==3.14.0
frozenlist==1.4.1
fsspec==2024.3.1
ftfy==6.2.0
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.63.0
huggingface-hub==0.23.0
hyper-connections==0.2.1
idna==3.7
importlib_metadata==7.1.0
Jinja2==3.1.3
local-attention==1.11.2
Markdown==3.6
MarkupSafe==2.1.5
modelcards==0.1.6
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.3
numpy==1.26.0
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
olefile==0.47
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
peft==0.7.1
performer-pytorch==1.1.3
pillow==10.3.0
platformdirs==4.2.1
protobuf==4.25.3
psutil==5.9.0
pyarrow==16.0.0
pyarrow-hotfix==0.6
pyg-lib==0.4.0+pt112cu113
pynvml==12.0.0
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.4.28
requests==2.31.0
safetensors==0.4.3
scipy==1.11.2
sentry-sdk==2.1.1
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy==1.13.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tokenizers==0.14.1
torch==1.12.1+cu113
torch-cluster==1.6.3
torch-geometric==2.6.1
torch-scatter==2.1.2
torch-sparse==0.6.18
torch-spline-conv==1.2.2
torchaudio==0.12.1+cu113
torchvision==0.13.1+cu113
tqdm==4.66.4
transformers==4.34.0
triton==3.2.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wandb==0.17.0
wcwidth==0.2.13
Werkzeug==3.0.2
xxhash==3.4.1
yarl==1.9.4
zipp==3.18.1
